Platformio reorganization
