+++ lib/Bridge/Bridge.h
