+++ sam/src/ReadLight.cpp.NEW
