+++ sam/src/sckAux.h
